backend/requirements.txt
fastapi
uvicorn
python-socketio[asgi]
openai
gTTS
python-dotenv
flask-cors
pandas
pinecone-client